todo-flask/requirements.txt
PyJWT~=2.6.0
Flask~=2.2.3
Flask-Cors==3.0.10
SQLAlchemy~=2.0.11
PyMySQL==1.0.3
pytest~=7.3.1